First working prototype (IPv4 DHCP relay works in both directions)
In this version of the relay, we assume that client requests are
received on a double tagged VLAN interface and server replies are
received on a single tagged VLAN interface.

Test scripts are added:

- test.sh (configures BNG and sets L3 config)
- cleanup_test.sh (removes BNG configuration)
- trace.sh (opens trace pipe for debugging)

The code for adding interface name is disabled because it makes the
verifier complain about the BPF program being too large.

To-do:
- Fix verifier issue to include interface name in option 82
- For server replies, erase option 82 instead of (partial) overwrite

Signed-off-by: Yoel Caspersen <yoel@kviknet.dk>