Improve readability of write_dhcp_option_82()
This makes several changes to improve the readability of write_dhcp_option_82():

The main change is in reworking the string helpers to be smaller and more
modular, and using a separate buffer for stringifying the VLAN tags before
copying them to the DHCP option field. Get rid of str_len() and the len
member of struct dev_name. To aid in this, the whole containing function is
made global (i.e., a separate BPF subprog) instead of the helpers.

In addition, the local variables are renamed to be more descriptive, and
the option struct is initialised using struct initialisation syntax.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>